requirements.txt
fastapi==0.110.0
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.25
mysql-connector-python==8.4.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
uvicorn==0.27.1
pytest==8.0.0
httpx==0.26.0
itsdangerous==2.1.2
argon2-cffi==23.1.0
tzdata==2024.1
